refactor(base_types): Move `AccessList` to base types
