+++ src/ethereum_test_base_types/__init__.py
@@ -18,7 +18,7 @@ from .base_types import (
     Wei,
     ZeroPaddedHexNumber,
 )
-from .composite_types import Account, Alloc, Storage, StorageRootType
+from .composite_types import AccessList, Account, Alloc, Storage, StorageRootType
 from .constants import (
     AddrAA,
     AddrBB,
@@ -35,6 +35,7 @@ from .pydantic import CamelModel
 from .reference_spec import ReferenceSpec
 
 __all__ = (
+    "AccessList",
     "Account",
     "AddrAA",
     "AddrBB",
