+++ src/ethereum_test_fixtures/state.py
@@ -6,10 +6,9 @@ from typing import ClassVar, List, Mapping, Sequence
 
 from pydantic import BaseModel, Field
 
-from ethereum_test_base_types import Address, Alloc, Bytes, Hash, ZeroPaddedHexNumber
+from ethereum_test_base_types import AccessList, Address, Alloc, Bytes, Hash, ZeroPaddedHexNumber
 from ethereum_test_exceptions import TransactionExceptionInstanceOrList
 from ethereum_test_types.types import (
-    AccessList,
     AuthorizationTupleGeneric,
     CamelModel,
     EnvironmentGeneric,
